+++ packages/backend/src/nest/storage/communityMetadata/communityMetadata.store.ts
@@ -64,17 +64,13 @@ export class CommunityMetadataStore extends KeyValueStoreBase<CommunityMetadata>
 
     this.store.events.on('replicated', async () => {
       logger('Replicated community metadata')
-      // @ts-expect-error - OrbitDB's type declaration of `load` lacks 'options'
-      // TODO: Is this necessary here?
-      await this.store.load({ fetchEntryTimeout: 15000 })
       const meta = this.getEntry()
       if (meta) {
         this.emit(StorageEvents.COMMUNITY_METADATA_STORED, meta)
       }
     })
 
-    // @ts-expect-error - OrbitDB's type declaration of `load` lacks 'options'
-    await this.store.load({ fetchEntryTimeout: 15000 })
+    await this.store.load()
     const meta = this.getEntry()
     if (meta) {
       this.emit(StorageEvents.COMMUNITY_METADATA_STORED, meta)