+++ packages/backend/src/nest/storage/userProfile/userProfile.store.ts
@@ -74,8 +74,7 @@ export class UserProfileStore extends KeyValueStoreBase<UserProfile> {
       })
     })
 
-    // @ts-expect-error - OrbitDB's type declaration of `load` lacks 'options'
-    await this.store.load({ fetchEntryTimeout: 15000 })
+    await this.store.load()
   }
 
   public getEntry(key: string): UserProfile {