+++ packages/mobile/src/store/init/startConnection/startConnection.saga.ts
@@ -1,5 +1,17 @@
 import { io } from 'socket.io-client'
-import { select, put, call, cancel, fork, takeEvery, FixedTask, delay, apply } from 'typed-redux-saga'
+import {
+  select,
+  put,
+  putResolve,
+  call,
+  cancel,
+  fork,
+  take,
+  takeLeading,
+  takeEvery,
+  FixedTask,
+  apply,
+} from 'typed-redux-saga'
 import { PayloadAction } from '@reduxjs/toolkit'
 import { socket as stateManager, Socket } from '@quiet/state-manager'
 import { encodeSecret } from '@quiet/common'
@@ -13,16 +25,6 @@ export function* startConnectionSaga(
 ): Generator {
   const isAlreadyConnected = yield* select(initSelectors.isWebsocketConnected)
   if (isAlreadyConnected) return
-
-  while (true) {
-    const isCryptoEngineInitialized = yield* select(initSelectors.isCryptoEngineInitialized)
-    console.log('WEBSOCKET', 'Waiting for crypto engine to initialize')
-    if (!isCryptoEngineInitialized) {
-      yield* delay(500)
-    } else {
-      break
-    }
-  }
 
   const { dataPort, socketIOSecret } = action.payload
 
@@ -49,17 +51,20 @@ export function* startConnectionSaga(
   })
   yield* fork(handleSocketLifecycleActions, socket, action.payload)
   // Handle opening/restoring connection
-  yield* takeEvery(initActions.setWebsocketConnected, setConnectedSaga, socket)
+  yield* takeLeading(initActions.setWebsocketConnected, setConnectedSaga, socket)
 }
 
 function* setConnectedSaga(socket: Socket): Generator {
+  console.log('Frontend is ready. Forking state-manager sagas and starting backend...')
+
   const task = yield* fork(stateManager.useIO, socket)
-  console.log('WEBSOCKET', 'Forking state-manager sagas', task)
-  // Handle suspending current connection
-  yield* takeEvery(initActions.suspendWebsocketConnection, cancelRootTaskSaga, task)
-  console.log('Frontend is ready. Starting backend...')
+
   // @ts-ignore - Why is this broken?
   yield* apply(socket, socket.emit, [SocketActionTypes.START])
+
+  // Handle suspending current connection
+  yield* take(initActions.suspendWebsocketConnection)
+  yield* call(cancelRootTaskSaga, task)
 }
 
 function* handleSocketLifecycleActions(socket: Socket, socketIOData: WebsocketConnectionPayload): Generator {
@@ -91,5 +96,5 @@ function subscribeSocketLifecycle(socket: Socket, socketIOData: WebsocketConnect
 function* cancelRootTaskSaga(task: FixedTask<Generator>): Generator {
   console.warn('Canceling root task', task.error())
   yield* cancel(task)
-  yield* put(initActions.canceledRootTask())
+  yield* putResolve(initActions.canceledRootTask())
 }