+++ packages/mobile/src/store/store.ts
@@ -71,5 +71,6 @@ export const store = configureStore({
 })
 
 export const persistor = persistStore(store, {}, () => {
+  console.log('Redux store is ready!')
   store.dispatch(initActions.setStoreReady())
 })