+++ packages/state-manager/src/sagas/identity/identity.master.saga.ts
@@ -1,5 +1,5 @@
 import { type Socket } from '../../types'
-import { all, takeEvery } from 'typed-redux-saga'
+import { all, takeEvery, cancelled } from 'typed-redux-saga'
 import { identityActions } from './identity.slice'
 import { registerUsernameSaga } from './registerUsername/registerUsername.saga'
 import { verifyJoinTimestampSaga } from './verifyJoinTimestamp/verifyJoinTimestamp.saga'
@@ -8,10 +8,18 @@ import { usersActions } from '../users/users.slice'
 import { updateCertificateSaga } from './updateCertificate/updateCertificate.saga'
 
 export function* identityMasterSaga(socket: Socket): Generator {
-  yield all([
-    takeEvery(identityActions.registerUsername.type, registerUsernameSaga, socket),
-    takeEvery(identityActions.verifyJoinTimestamp.type, verifyJoinTimestampSaga),
-    takeEvery(identityActions.saveUserCsr.type, saveUserCsrSaga, socket),
-    takeEvery(usersActions.responseSendCertificates.type, updateCertificateSaga),
-  ])
+  console.log('identityMasterSaga starting')
+  try {
+    yield all([
+      takeEvery(identityActions.registerUsername.type, registerUsernameSaga, socket),
+      takeEvery(identityActions.verifyJoinTimestamp.type, verifyJoinTimestampSaga),
+      takeEvery(identityActions.saveUserCsr.type, saveUserCsrSaga, socket),
+      takeEvery(usersActions.responseSendCertificates.type, updateCertificateSaga),
+    ])
+  } finally {
+    console.log('identityMasterSaga stopping')
+    if (yield cancelled()) {
+      console.log('identityMasterSaga cancelled')
+    }
+  }
 }