+++ packages/desktop/CHANGELOG.md
@@ -6,6 +6,9 @@
 
 # Fixes:
 
+* Update github workflows for PR gating ([#2487](https://github.com/TryQuiet/quiet/issues/2487))
+* Don't create duplicate CSRs when joining a community under certain circumstances ([#2321](https://github.com/TryQuiet/quiet/issues/2321))
+
 [2.2.0]
 
 # New features:
@@ -19,9 +22,9 @@
 
 # Fixes:
 
-* Allow JPEG and GIF files as profile photos ([#2332](https://github.com/TryQuiet/quiet/issues/2332))
 * Fix issues with recreating general channel when deleted while offline ([#2334](https://github.com/TryQuiet/quiet/issues/2334))
 * Fix package.json license inconsistency
+* Fixes issue with reconnecting to peers on resume on iOS ([#2424](https://github.com/TryQuiet/quiet/issues/2424))
 
 [2.1.2]